Time: 3 ms (70.39%), Space: 5.8 MB (73.31%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-     // return sqrt(x);
+        // return sqrt(x);
         // if(x<2) return x;
         // auto i=1;
         // while(x/i>i)
@@ -9,7 +9,7 @@ public:
         //   i++;
         // }
         // return x/i;
-      long long s=0,e=x,ans,mid;//long long due to some of test cases overflows integer limit.
+      long s=0,e=x,ans,mid;//long long due to some of test cases overflows integer limit.
         while(s<=e){             
             mid=(s+e)/2;
             if(mid*mid==x) return mid;//if the 'mid' value ever gives the result, we simply return it.
